refacto home page form with hook and state
